Extract code related to the stats of the real errors from real.py to real_errors_stats.py
